cloud/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0

